+++ src/main/java/me/ajlane/neo4j/Neo4jAgroSuccessLct.java
@@ -19,11 +19,11 @@ public class Neo4jAgroSuccessLct implements LandCoverTypeEco, LandCoverTypeAnthr
   public Neo4jAgroSuccessLct(Node node) {
     this.code = (String) node.getProperty("code");
     this.descr = (String) node.getProperty("description");
-    this.num = (new Long((long) node.getProperty("num"))).intValue();
+    this.num = (Long.valueOf((long) node.getProperty("num"))).intValue();
     this.isMatureVegetation = (boolean) node.getProperty("is_mature_vegetation");
-    this.fertility = (new Long((long) node.getProperty("fertility"))).intValue();
+    this.fertility = (Long.valueOf((long) node.getProperty("fertility"))).intValue();
     this.landCoverConversionCost =
-        (new Long((long) node.getProperty("land_cover_conversion_cost"))).intValue();
+        (Long.valueOf((long) node.getProperty("land_cover_conversion_cost"))).intValue();
   }
 
   @Override
@@ -31,22 +31,27 @@ public class Neo4jAgroSuccessLct implements LandCoverTypeEco, LandCoverTypeAnthr
     return code;
   }
 
+  @Override
   public String getDescr() {
     return descr;
   }
 
+  @Override
   public int getNum() {
     return num;
   }
 
+  @Override
   public boolean getIsMatureVegetation() {
     return isMatureVegetation;
   }
 
+  @Override
   public int getFertility() {
     return fertility;
   }
 
+  @Override
   public int getLandCoverConversionCost() {
     return landCoverConversionCost;
   }