+++ src/test/java/repast/model/agrosuccess/AgroSuccessEnvrIntegrationTest.java
@@ -60,7 +60,7 @@ public class AgroSuccessEnvrIntegrationTest {
     // default value in parameters.xml
     params.setValue("nTicks", 100);
 
-    params.setValue("useDummyData", new Boolean(true));
+    params.setValue("useDummyData", Boolean.TRUE);
 
     // Boost climateIgnitionScalingParameter to ensure fires take place for testing purposes
     params.setValue("climateIgnitionScalingParam", 24);